Use tempita from cython 0.25.2
with minimal modification
